Updated LinAlg library to support NxM matrices instead of just square 3x3. Far from a full generalization yet

+++ Code/main/headers/Config.h
@@ -8,16 +8,16 @@
 /*---Controller---*/
 
 //Controller gains:
-extern const float KpX, KpY, KpZ;
-extern const float KdX, KdY, KdZ;
-extern const float KiX, KiY, KiZ;
+extern const double KpX, KpY, KpZ;
+extern const double KdX, KdY, KdZ;
+extern const double KiX, KiY, KiZ;
 //Throttle limits:
-extern const float TMax; // This is the maximum pulse width the PD controller is allowed to send to the motors
-extern const float TMin; // This is the minimum pulse width the PD controller is allowed to send to the motors 
-extern const float TCutoff; // If T is below this, the motors will be switched off
-extern const float TPowerOff; //This is a safe value the motors will be initialized with
+extern const double TMax; // This is the maximum pulse width the PD controller is allowed to send to the motors
+extern const double TMin; // This is the minimum pulse width the PD controller is allowed to send to the motors 
+extern const double TCutoff; // If T is below this, the motors will be switched off
+extern const double TPowerOff; //This is a safe value the motors will be initialized with
 //Integrator wind-up limits:
-extern const float WindUpX, WindupY, WindUpZ;
+extern const double WindUpX, WindupY, WindUpZ;
 
 //Arm threshold for the four radio channels
 extern const uint32_t ArmCh0, ArmCh1, ArmCh2, ArmCh3;
@@ -26,15 +26,15 @@ extern const uint32_t ArmCh0, ArmCh1, ArmCh2, ArmCh3;
 //#define RATE_ROLL_CONTROL //Not implemented
 //#define RATE_PITCH_CONTROL //Not implemented
 #define RATE_YAW_CONTROL //Comment this out to control yaw angle instead
-extern const float RollRate, PitchRate, YawRate;
+extern const double RollRate, PitchRate, YawRate;
 
 /*---Estimator---*/
 
 //Estimator gains:
-extern const float KpXe, KpYe, KpZe; //For the complementary filter
-extern const float KiXe, KiYe, KiZe; //For gyro-bias estimation
+extern const double KpXe, KpYe, KpZe; //For the complementary filter
+extern const double KiXe, KiYe, KiZe; //For gyro-bias estimation
 
 //Gyroscope LP-filter time constant. Note that this is done on the RP2350, not the IMU
-extern const float Lambda;
+extern const double Lambda;
 
 #endif
