+++ Code/main/src/Config.c
@@ -3,28 +3,28 @@
 /*---Controller---*/
 
 //Controller gains:
-const float KpX = 1500.f;
-const float KpY = 1500.f;
-const float KpZ = 1500.f;
+const double KpX = 1500.0;
+const double KpY = 1500.0;
+const double KpZ = 1500.0;
 
-const float KdX = 150.f;
-const float KdY = 150.f;
-const float KdZ = 0.f;
+const double KdX = 150.0;
+const double KdY = 150.0;
+const double KdZ = 0.0;
 
-const float KiX = 0.f; //450.f;
-const float KiY = 0.f; //450.f;
-const float KiZ = 0.f;
+const double KiX = 0.0; //450.0;
+const double KiY = 0.0; //450.0;
+const double KiZ = 0.0;
 
 //Integrator wind-up limits:
-const float WindUpX = 100.f;
-const float WindupY = 100.f;
-const float WindUpZ = 100.f;
+const double WindUpX = 100.0;
+const double WindupY = 100.0;
+const double WindUpZ = 100.0;
 
 //Throttle limits:
-const float TMax = 1920.f;
-const float TMin = 1050.f;
-const float TCutoff = 980.f;
-const float TPowerOff = 0.f;
+const double TMax = 1920.0;
+const double TMin = 1050.0;
+const double TCutoff = 980.0;
+const double TPowerOff = 0.0;
 
 //Arm threshold for the four radio channels
 const uint32_t ArmCh0 = 1200; //Yaw
@@ -33,23 +33,23 @@ const uint32_t ArmCh2 = 1200; //Pitch
 const uint32_t ArmCh3 = 1200; //Roll
 
 //Controller behavior selection
-const float RollRate = 1.f; //Not implemented
-const float PitchRate = 1.f; //Not implemented
-const float YawRate = 1.f;
+const double RollRate = 1.0; //Not implemented
+const double PitchRate = 1.0; //Not implemented
+const double YawRate = 1.0;
 
 
 
 /*---Estimator---*/
 
 //For the complementary filter
-const float KpXe = 0.5f;
-const float KpYe = 0.5f; 
-const float KpZe = 0.0f; //Set to 0 until magnetometer is introduced
+const double KpXe = 0.5;
+const double KpYe = 0.5; 
+const double KpZe = 0.0; //Set to 0 until magnetometer is introduced
 
 //For gyro-bias estimation
-const float KiXe = 0.15f;
-const float KiYe = 0.15f; 
-const float KiZe = 0.0f; //Set to 0 until magnetometer is introduced
+const double KiXe = 0.15;
+const double KiYe = 0.15; 
+const double KiZe = 0.0f; //Set to 0 until magnetometer is introduced
 
 //Gyroscope LP-filter time constant
-const float Lambda = 0.0107f;
+const double Lambda = 0.0107;
